+++ packages/kit/src/runtime/client/types.d.ts
@@ -5,6 +5,7 @@ export type NavigationInfo = {
 	id: string;
 	routes: CSRRoute[];
 	path: string;
+	decoded_path: string;
 	query: URLSearchParams;
 };
 
@@ -28,10 +29,10 @@ export type BranchNode = {
 		path: boolean;
 		query: boolean;
 		session: boolean;
-		context: boolean;
+		stuff: boolean;
 		dependencies: string[];
 	};
-	context: Record<string, any>;
+	stuff: Record<string, any>;
 };
 
 export type NavigationState = {