+++ packages/kit/src/runtime/server/page/types.d.ts
@@ -3,7 +3,8 @@ import { NormalizedLoadOutput, SSRNode } from 'types/internal';
 export type Loaded = {
 	node: SSRNode;
 	loaded: NormalizedLoadOutput;
-	context: Record<string, any>;
+	stuff: Record<string, any>;
 	fetched: Array<{ url: string; body: string; json: string }>;
+	set_cookie_headers: string[];
 	uses_credentials: boolean;
 };