+++ packages/kit/src/runtime/utils.js
@@ -1,7 +1,0 @@
-/**
- * @param {unknown} err
- * @return {Error}
- */
-export function coalesce_to_error(err) {
-	return err instanceof Error ? err : new Error(JSON.stringify(err));
-}