+++ packages/kit/test/apps/basics/src/routes/load/stuff/__layout.svelte
@@ -1,9 +1,9 @@
 <script context="module">
 	/** @type {import('@sveltejs/kit').Load} */
-	export async function load({ page }) {
+	export async function load() {
 		return {
-			context: {
-				y: page.params.y
+			stuff: {
+				message: 'original'
 			}
 		};
 	}