+++ packages/kit/test/apps/basics/src/routes/nested-layout/_tests.js
@@ -32,4 +32,16 @@ export default function (test) {
 			assert.equal(await page.textContent('h1'), '500');
 		}
 	);
+
+	test(
+		'renders deeply-nested errors in the right layout',
+		'/nested-layout/foo/bar/nope',
+		async ({ page }) => {
+			assert.equal(await page.textContent('footer'), 'Custom layout');
+			assert.ok(await page.evaluate(() => document.querySelector('p#nested')));
+			assert.ok(await page.evaluate(() => document.querySelector('p#nested-foo')));
+			assert.ok(await page.evaluate(() => document.querySelector('p#nested-bar')));
+			assert.equal(await page.textContent('#nested-error-message'), 'error.message: nope');
+		}
+	);
 }