+++ packages/kit/types/config.d.ts
@@ -1,5 +1,5 @@
 import { UserConfig as ViteConfig } from 'vite';
-import { JSONString, RecursiveRequired } from './helper';
+import { RecursiveRequired } from './helper';
 import { Logger, TrailingSlash } from './internal';
 
 export interface AdapterUtils {
@@ -51,15 +51,8 @@ export interface Config {
 		package?: {
 			dir?: string;
 			emitTypes?: boolean;
-			exports?: {
-				include?: string[];
-				exclude?: string[];
-			};
-			files?: {
-				include?: string[];
-				exclude?: string[];
-			};
-			override?: Record<string, JSONString>;
+			exports?(filepath: string): boolean;
+			files?(filepath: string): boolean;
 		};
 		paths?: {
 			assets?: string;
@@ -68,12 +61,12 @@ export interface Config {
 		prerender?: {
 			crawl?: boolean;
 			enabled?: boolean;
+			entries?: string[];
 			onError?: PrerenderOnErrorValue;
-			pages?: string[];
 		};
 		router?: boolean;
 		serviceWorker?: {
-			exclude?: string[];
+			files?(filepath: string): boolean;
 		};
 		ssr?: boolean;
 		target?: string;